remove dependency to gnAsteroid/gno, depend on gnolang/gno instead. Integrate gnoweb1's gnoweb.go
Initial plan was to depend on gnoweb 2.0. After digging a little bit,
it would take a lot of time (all components, views would need to be
rewritten). Besides, gnoweb 1.0, with its md files and gorilla, is still a
very fine piece of software, if gno can work with both gnoweb 2.0 and
gnAsteroid it may be good for gno.